+++ index.html
@@ -231,6 +231,8 @@
   <template id="messages">
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
-  <script src="js/main.js"></script>
+  <script type="module" src="js/data.js"></script>
+  <script type="module" src="js/util.js"></script>
+  <script type="module" src="js/main.js"></script>
 </body>
 </html>
