+++ js/util.js
@@ -1,0 +1,26 @@
+export const getRandomIntegerFromRange = (start, end) =>
+  Math.ceil(Math.random() * (end - start + 1)) + start - 1;
+
+export const getRandomInteger = () => {
+  let lastGeneratedId = 0;
+
+  return function () {
+    lastGeneratedId += 1;
+    return lastGeneratedId;
+  };
+};
+
+export const createRandomIntFromRangeGenerator = (min, max) => {
+  const previousValues = [];
+
+  return function () {
+    let currentValue = getRandomIntegerFromRange(min, max);
+
+    while (previousValues.includes(currentValue)) {
+      currentValue = getRandomIntegerFromRange(min, max);
+    }
+    previousValues.push(currentValue);
+
+    return currentValue;
+  };
+};
